Allow to install latest available mesos version
This patch allows wrapper cookbooks to choose to upgrade to the latest
version available in repositories.

This behavior can be useful for test nodes where you always want latest
version of rpm/deb.
For instance, a cluster could have a few percent of slaves using latest
version of mesos rpm to allow users to test some features in advance.

Change-Id: Idcf99ab4a3925401d52ee511960e0c6f28001214